Upgrade serializer to new construct
This provides with a few options, as summarized in the docstrings.
